Added policy method so hoom cant be deleted

+++ app/Http/Controllers/Cooperation/Admin/SuperAdmin/Cooperation/CooperationController.php
@@ -58,6 +58,8 @@ class CooperationController extends Controller
 
     public function destroy(Cooperation $cooperation, Cooperation $cooperationToDestroy)
     {
+
+        $this->authorize('delete', $cooperationToDestroy);
         $cooperationToDestroy->steps()->detach();
 
         $exampleBuildings = $cooperationToDestroy->exampleBuildings;
